requirements.txt
asgiref==3.2.10
astroid==2.4.2
certifi==2020.6.20
chardet==3.0.4
Django==3.1.2
django-elasticsearch-dsl==7.1.4
django-oauth==1.1
djangorestframework==3.12.1
elasticsearch==7.9.1
elasticsearch-dsl==7.3.0
idna==2.10
isort==5.5.4
lazy-object-proxy==1.5.1
mccabe==0.6.1
oauthlib==3.1.0
Pillow==7.2.0
psycopg2-binary==2.8.6
Pygments==2.7.1
pylint==2.6.0
python-dateutil==2.8.1
python-decouple==3.3
pytz==2020.1
requests==2.24.0
requests-oauthlib==1.3.0
six==1.15.0
sqlparse==0.3.1
toml==0.10.1
typed-ast==1.4.1
urllib3==1.25.10
wrapt==1.12.1
